LayoutTests/media/media-fragments/TC0050-TC0059-expected.txt


Title: TC0051
Fragment: 't=3&'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0052
Fragment: 'u=12&t=3'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0053
Fragment: 't=foo:7&t=npt:3'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0054
Fragment: '&&=&=tom&jerry=&t=3&t=meow:0#'
Comment: After processing name-value pairs, this appears to be equivalent to #t=3.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0055
Fragment: 't=7&t=3'
Comment: When a fragment dimensions occurs multiple times, only the last occurrence of that dimension is interpreted.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

Title: TC0058
Fragment: 'T=3,7'
Comment: UA does not identify this as a media fragment. The entire media resource is played.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '0') OK
EXPECTED (video.initialTime == '0') OK

Title: TC0059
Fragment: 't=smpte:00:00:01&t=npt:3'
Comment: When a fragment dimensions occurs multiple times, only the last occurrence of that dimension is interpreted.
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
EXPECTED (video.initialTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK

END OF TEST
